requirements.txt
playwright==1.41.2
aiofiles==23.2.1
python-dotenv==1.0.0 